+++ components/layout/page-navigation/PageNavigation.tsx
@@ -3,6 +3,8 @@ import { FC } from 'react'
 import DefaultNoLinkNavigation from '@/components/layout/page-navigation/DefaultNoLinkNavigation'
 import PageNavigationLink from '@/components/layout/page-navigation/PageNavigationLink'
 
+import { DATA_TEST_IDS } from '@/__tests__/playwright/lib/utils/constants/ids/dataTestIds'
+
 import { NavigationDirectionEnum } from '@/lib/utils/interfaces/enums'
 
 type PageNavigationProps = {
@@ -26,32 +28,34 @@ const PageNavigation: FC<PageNavigationProps> = ({
   const hasNextLink = linkNext && nameNext
 
   return (
-    <div className="container mx-auto mt-20 max-w-screen-xl">
-      <div className={`group flex flex-col gap-4 md:flex-row ${hasPreviousLink ? 'justify-between' : 'justify-end'}`}>
-        {hasPreviousLink ? (
-          <PageNavigationLink
-            href={linkPrevious}
-            text={namePrevious}
-            direction={NavigationDirectionEnum.Left}
-            dataTestId={dataTestIdPrevious}
-            justify="start"
-          />
-        ) : (
-          <DefaultNoLinkNavigation arrowDirection={NavigationDirectionEnum.Left} />
-        )}
-        {hasNextLink ? (
-          <PageNavigationLink
-            href={linkNext}
-            text={nameNext}
-            direction={NavigationDirectionEnum.Right}
-            dataTestId={dataTestIdNext}
-            justify="end"
-          />
-        ) : (
-          <DefaultNoLinkNavigation arrowDirection={NavigationDirectionEnum.Right} />
-        )}
+    <nav data-testid={DATA_TEST_IDS.navigation.END_OF_PAGE}>
+      <div className="container mx-auto mt-20 max-w-screen-xl">
+        <div className={`group flex flex-col gap-4 md:flex-row ${hasPreviousLink ? 'justify-between' : 'justify-end'}`}>
+          {hasPreviousLink ? (
+            <PageNavigationLink
+              href={linkPrevious}
+              text={namePrevious}
+              direction={NavigationDirectionEnum.Left}
+              dataTestId={dataTestIdPrevious}
+              justify="start"
+            />
+          ) : (
+            <DefaultNoLinkNavigation arrowDirection={NavigationDirectionEnum.Left} />
+          )}
+          {hasNextLink ? (
+            <PageNavigationLink
+              href={linkNext}
+              text={nameNext}
+              direction={NavigationDirectionEnum.Right}
+              dataTestId={dataTestIdNext}
+              justify="end"
+            />
+          ) : (
+            <DefaultNoLinkNavigation arrowDirection={NavigationDirectionEnum.Right} />
+          )}
+        </div>
       </div>
-    </div>
+    </nav>
   )
 }
